% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{chain.mixing}
\alias{chain.mixing}
\title{Chain mixing plot for piecewise exponential model}
\usage{
chain.mixing(object)
}
\arguments{
\item{object}{an object of class "changepoint".}
}
\value{
plot with the number of change-points at each simulation, coloured by chain number
}
\description{
Chain mixing plot for piecewise exponential model
}
\examples{
\dontrun{chain.mixing(Collapsing_Model)}
}
