% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{compare.surv.mods}
\alias{compare.surv.mods}
\title{Comparing Piecewise Exponential model with other Parametric models}
\usage{
compare.surv.mods(object, max_predict = 10)
}
\arguments{
\item{object}{of class "changepoint".}

\item{max_predict}{maximum survival time to be predicted for the survival models. Default is 10, however, depending on the timescale this should be changed.}
}
\value{
A list of with the following items:
\itemize{
\item \strong{model.comp}: A dataframe with the PML and WAIC for the piecewise exponential model and the six parametric models fitted by JAGS.
\item \strong{jags.models}: A list containing the posterior simulations of the 6 JAGS models (fit using the R2jags:jags function).
\item \strong{plot_Surv_all}: A ggplot with the posterior mean survival probabilities for the time specified by max_predict.
}
}
\description{
Compares the piecewise exponential model with 6 other standard parameteric models (see \code{\link[=fit_surv_models]{fit_surv_models()}}). Note that exponential is a special case of the piecewise exponential, however, it is refit in JAGS to highlight the difference in statistical fit.
This functions computes the individual log-likelihood for the piecewise exponential model (see \code{\link[=get.loglik]{get.loglik()}}) and compares the Widely Applicable Information Criterion (using the \code{\link[loo:loo-package]{loo::loo-package}}) and Pseudo-Marginal Likelihood (PML) with the other standard parametric models.
}
\examples{
\dontrun{
mod.comp <-compare.surv.mods(Collapsing_Model, df)}
}
