% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{digitise}
\alias{digitise}
\title{Digitise KM curves}
\usage{
digitise(
  surv_inp,
  nrisk_inp,
  km_output = "KMdata.txt",
  ipd_output = "IPDdata.txt"
)
}
\arguments{
\item{ipd_output}{}
}
\description{
Digitise KM curves
}
