% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{gen_piece_df}
\alias{gen_piece_df}
\title{Piecewise Exponential Simulations
Generate piecewise exponential observations.}
\usage{
gen_piece_df(n_obs, n_events_req, num.breaks, rate, t_change, max_time = Inf)
}
\arguments{
\item{n_obs}{number of observations}

\item{n_events_req}{(approximate) number of events contained within the sample}

\item{num.breaks}{number of change-points, NA if no changepoints}

\item{rate}{vector of hazard rate}

\item{t_change}{vector of change-point times}

\item{max_time}{maximum time after which all observations are censored.}
}
\value{
a dataframe with three columns, the time of the events (time_event), time which is the minimum of the cenorsing time and event time. status is an indicator which is 0 censored observation, or 1 if event.
}
\description{
Piecewise Exponential Simulations
Generate piecewise exponential observations.
}
\examples{
set.seed(123)
n_obs =300
n_events_req=300
max_time =  2
rate = c(0.75,0.25)
t_change =1
df <- gen_piece_df(n_obs = n_obs,n_events_req = n_events_req,
                  num.breaks = length(t_change),rate = rate ,
                  t_change = t_change, max_time = max_time)
}
