% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get.loglik}
\alias{get.loglik}
\title{Pointwise log-likelihood for the change-point model}
\usage{
get.loglik(object, chng.num = "all")
}
\arguments{
\item{object}{of class "changepoint".}

\item{chng.numvalue}{indicating the changepoint model to plotted, default is "all", in which all posterior simulations will be used.}
}
\value{
a matrix of size nSims by n_obs (number of simulations and number of observations respectively).
}
\description{
Provides the pointwise log-likelihood contribution for each datapoint and simulation which is required to compute the Pseudo-Marginal Likelihood (PML) and Widely Applicable Information Criterion (WAIC).
}
\examples{
\dontrun{
Can take a considerable time to evaulate.
log.lik.df <- get.loglik(Collapsing_Model)
}
}
