% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{plot.changepoint}
\alias{plot.changepoint}
\title{Plot functions for change-point models}
\usage{
\method{plot}{changepoint}(
  object,
  type = "survival",
  chng.num = "all",
  add.km = T,
  max_predict = 10,
  add.post = T,
  alpha.pos = NULL,
  t_pred = NULL,
  final_chng_only = F,
  col_km = "black",
  km_risk = NULL,
  ...
)
}
\arguments{
\item{object}{an object of class "changepoint".}

\item{type}{the type of plot to be drawn, default is the survival function. Also can plot the hazard function with "hazard".}

\item{chng.num}{value indicating the changepoint model to plotted, default is "all", in which all posterior simulations will be used.}

\item{add.km}{indicator to add Kaplan Meier curve. Only applicable to survival function. Default is true.}

\item{max_predict}{maximum time to be plotted. Default is 10, however, depending on the timescale this should be changed.}

\item{add.post}{indicator whether to plot the posterior simulations (a random sample of 500) for the survival function. Default is true.}
}
\value{
a ggplot object
}
\description{
Plot functions for change-point models
}
\examples{
\dontrun{
plot(Collapsing_Model, add.post = F)
plot(Collapsing_Model, type = "hazard")}
}
