% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{print.changepoint}
\alias{print.changepoint}
\title{Printing of Changepoint model objects}
\usage{
\method{print}{changepoint}(object, chng.num = NULL, digits = min(3L, getOption("digits")))
}
\arguments{
\item{object}{of class "changepoint".}

\item{chng.num}{indicating the change-point model to summarized in terms of change-point(s) and hazards, if ommited the model with the highest posterior probability is presented.}

\item{digits}{number of digits to be printed.}
}
\description{
Printing of Changepoint model objects
}
\examples{
\notrun{
Print a summary of the one change-point model
print(Collapsing_Model, mod = 1)
Print a summary of the model with the highest posterior probability
print(Collapsing_Model)
}
}
