% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapsing_model.R
\name{collapsing.model}
\alias{collapsing.model}
\title{Fitting Bayesian Piecewise Exponential Models}
\usage{
collapsing.model(
  df,
  n.iter = 5750,
  burn_in = 750,
  n.chains = 2,
  max.num.breaks = 6,
  n.thin = 1,
  lambda.prior = 1,
  alpha.hyper = 1,
  beta.hyper1 = 1,
  beta.hyper2 = 1
)
}
\arguments{
\item{df}{A dataframe with survival times and status indicator.}

\item{n.iter}{number of total iterations per chain excluding burn in (default 10750).}

\item{burn_in}{length of burn in, i.e. number of iterations to discard at the beginning of each chain (default 750).}

\item{n.chains}{number of Markov chains (default 3).}

\item{max.num.breaks}{maximum number of allowed change-points (default 6).}

\item{n.thin}{thinning rate. Must be a positive integer. Set n.thin > 1 to save memory and computation time if n.iter is large (default 1).}

\item{lambda.prior}{rate parameter for the Poisson prior for number of change-points (default 1).}

\item{alpha.hyper}{value of shape parameter for the gamma prior for the hazard (\eqn{\lambda}). Should always be set to 1. As noted there is a gamma prior on \eqn{\lambda} but also a gamma prior (hyperprior) on the rate parameter of this prior.}

\item{beta.hyper1}{first hyperprior (shape) for the rate parameter of the gamma prior for \eqn{\lambda}. Should be set to 1.}

\item{beta.hyper2}{first hyperprior (rate) for the rate parameter of the gamma prior for \eqn{\lambda}. To insure that inferences are the same when different timescales (i.e. months vs years) are chosen, this hyperprior must be changed. If the analysis is in years this value should be equal to 1, if in months it should be 1/12 and if in days should be 1/365.}
}
\value{
A list of the class "changepoint" with the following items:
\itemize{
\item \strong{k.stacked}: Matrix of change-point indices. Number refers to the number of events.
\item \strong{changepoint}: Matrix of change-point indices in terms of time (rather than number of events).
\item \strong{lambda}: Matrix with post-hoc estimation of the hazards.
\item \strong{prob.changepoint}: Posterior probabilities of changepoint models.
\item \strong{k}: k.stacked but in array form with each slice indicating a chain.
\item \strong{beta.array} Posterior samples of the rate of the gamma prior for the (\eqn{\lambda}). Note that this is the rate of the (lower level) prior and not the hyperprior.
}
}
\description{
Implementation of piecewise exponential models typically assume that the number of a location of change-points is known. In many cases both of these quantities are unknown. This function estimates the posterior probability of change-point numbers and locations.
}
\examples{
\dontrun{Collapsing_Model <- collapsing.model(df,
                                                        n.iter = 5000,
                                                        burn_in = 750,
                                                        n.chains = 2,
                                                        alpha.hyper = 1,
                                                        beta.hyper1 = 1,
                                                        beta.hyper2 = 1)}
}
